server/requirements.txt
Flask
Flask-CORS
fpdf
python-docx
Pillow
